infra/scripts/data_scripts/requirements_load_tables.txt
azure-identity==1.12.0  # For DefaultAzureCredential to acquire access tokens
psycopg2-binary==2.9.6  # For connecting to PostgreSQL
pandas==2.1.0           # For reading and processing CSV files. Other versions: 2.2.2, 1.5.3

